requirements.txt
fastapi==0.115.0
pydantic==2.9.2
python-dotenv==1.0.1
motor==3.6.0
starlette==0.40.0

# Dev tools
uvicorn==0.30.6
python-jose[cryptography]==3.5.0
